requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
sentence-transformers
langchain_huggingface
faiss-cpu